refactor(login): clean up + CI passing before PR
