+++ lib/main.dart
@@ -1,7 +1,7 @@
 import "package:firebase_core/firebase_core.dart";
 import "package:flutter/material.dart";
 import "package:hooks_riverpod/hooks_riverpod.dart";
-import "package:proxima/utils/firebase_options.dart";
+import "package:proxima/utils/firebase/firebase_options.dart";
 import "package:proxima/views/navigation/routes.dart";
 
 void main() async {
