+++ lib/views/pages/login_page.dart
@@ -1,6 +1,6 @@
 import "package:flutter/material.dart";
 import "package:hooks_riverpod/hooks_riverpod.dart";
-import "package:proxima/viewmodels/login_viewmodel.dart";
+import "package:proxima/viewmodels/login_view_model.dart";
 import "package:proxima/views/navigation/routes.dart";
 
 class LoginPage extends HookConsumerWidget {
